requirements.txt
flask
flask_cors
Gunicorn
transformers
torch
nltk
python-dotenv
datasets
numpy
requests